Add test models, web to login and logout
