+++ apps/account/tests/__init__.py
@@ -1,1 +1,2 @@
-from .authentication.auth import SignInTest, SignInViewTest
+from apps.account.tests.models import SignInTest
+from apps.account.tests.web import SignInViewTest, LogoutTest
